requirements.txt
fastapi
uvicorn
sentence-transformers
transformers
torch
Pillow
yt-dlp
chromadb
python-dotenv
openai
